Corrected wrong attribute name in  CustomMDP.py.  Added correct CustomMDP class instantiation and output testing in main.py. See description for current error.
CURRENT ERROR: "our_mdp.states" does not have the correct states.  They are in a dictionary for some reason instead of a list, and also there are random characters included in the list of states.  If you look in MDP.py and see how our mdp is getting the list of states, you will find:

self.states = states or self.get_states_from_transitions(transitions)

---------------
This shows that the states are being extracted from our transitions somehow.  Possibly need to look and make sure transitions are in correct format to be extracted by this function.